some other module reorganization
