+++ zengine/src/platform/system.rs
@@ -4,10 +4,10 @@ use crate::core::system::System;
 use crate::core::system::Write;
 use crate::core::Store;
 use crate::core::Trans;
-use crate::event::controller::{ControllerButton, Which};
+use crate::device::controller::{ControllerButton, Which};
+use crate::device::keyboard::Key;
+use crate::device::mouse::MouseButton;
 use crate::event::input::{Axis, Input, InputEvent};
-use crate::event::keyboard::Key;
-use crate::event::mouse::MouseButton;
 use crate::event::EventStream;
 use log::{info, trace};
 use sdl2::event::Event;
